// packages/cli/templates/default/convex/_generated/dataModel.d.ts
export type DataModel = {
// ...
  threads: {
    document: {
      agentId: string;
      createdAt: number;
      metadata?: any;
      name?: string;
      projectId?: Id<"projects">;
      updatedAt: number;
      userId?: string;
      _id: Id<"threads">;
      _creationTime: number;
    };
    fieldPaths:
      | "_creationTime"
      | "_id"
      | "agentId"
      | "createdAt"
      | "metadata"
      | "name"
      | "projectId"
      | "updatedAt"
      | "userId";
    indexes: {
      by_id: ["_id"];
      by_creation_time: ["_creationTime"];
      byAgentId: ["agentId", "_creationTime"];
      byProjectId: ["projectId", "_creationTime"];
      byUserId: ["userId", "_creationTime"];
    };
    searchIndexes: {};
    vectorIndexes: {};
  };
// ...
};

/**
 * The names of all of your Convex tables.
 */
export type TableNames = TableNamesInDataModel<DataModel>;
// ...
export type Id<TableName extends TableNames | SystemTableNames> =
  GenericId<TableName>;